credit-check: have config just be config to avoid pickling issues but in a backwards compatible way
